adding more tests

+++ src/main/java/com/jf/sc2022/dto/PaymentDTO.java
@@ -7,7 +7,7 @@ import lombok.Value;
 @Value
 public class PaymentDTO {
     long   id;
-    long   amount;
+    double amount;
     String email;
     String chargeId;
 }
